added assessment result styles to seperate file
